✨ differentiate server action vs sync server action
