+++ packages/atom.io/realtime-client/src/index.ts
@@ -5,3 +5,4 @@ export * from "./pull-mutable"
 export * from "./pull-mutable-family-member"
 export * from "./push"
 export * from "./server-action"
+export * from "./sync-server-action"
