+++ packages/atom.io/realtime-react/src/index.ts
@@ -5,3 +5,4 @@ export * from "./use-pull-mutable"
 export * from "./use-pull-mutable-family-member"
 export * from "./use-push"
 export * from "./use-server-action"
+export * from "./use-sync-server-action"
