+++ packages/atom.io/realtime-react/src/use-sync-server-action.ts
@@ -1,17 +1,19 @@
 import * as AtomIO from "atom.io"
-import { StoreContext } from "atom.io/react"
+import { StoreContext, useO } from "atom.io/react"
 import * as RTC from "atom.io/realtime-client"
 import * as React from "react"
 
 import { useRealtimeService } from "./use-realtime-service"
 
-export function useServerAction<ƒ extends AtomIO.ƒn>(
+export function useSyncServerAction<ƒ extends AtomIO.ƒn>(
 	token: AtomIO.TransactionToken<ƒ>,
 ): (...parameters: Parameters<ƒ>) => ReturnType<ƒ> {
 	const store = React.useContext(StoreContext)
+	const updateQueueState = AtomIO.findState(RTC.updateQueueAtoms, token)
+	const updateQueue = useO(updateQueueState)
 
 	useRealtimeService(`tx:${token.key}`, (socket) =>
-		RTC.serverAction(token, socket, store),
+		RTC.syncServerAction(token, socket, updateQueue, store),
 	)
 	return AtomIO.runTransaction(token, store)
 }
